requirements.txt
anthropic>=0.25.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.29.0
torch>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
